base: merge gpu2/e2/rdg2 root classes into RefCountedRuntimeType
gpu2::RootEntity, e2::Thing, and rdg2::Entity were three structurally
identical "RefCounter + RuntimeType + id + name" base classes. Merge their
implementation into GN::RefCountedRuntimeType<ID> in base/rtti.h — a pure
mixin template that registers no runtime type and owns no static data, so
it is DLL-safe without explicit-instantiation tricks.

Concrete non-templated root classes derive from the mixin, register as the
root of their runtime-type chain, and define GN_API auto-id constructors
implemented inside GNcore so generated ids stay process-unique even across
DLL boundaries:

- GN::RCRT64 (Being<int64_t>): used directly by gpu2 (id type changes from
  uint64_t to int64_t) and by e2, which constructs through the caller-
  supplied-id ctor with universe-generated ids.
- rdg2::Entity (RefCountedRuntimeType<NeverOverflowingCounter>): keeps its
  128-bit id and GN_API ctor in artifact.cpp; a fmt formatter for the
  counter supports the shared destruction trace.

e2 cleanup rides along: thing.h renamed to e2.h; the Thing and
UniqueIdentifier typedefs are removed (RCRT64/int64_t used directly); the
macro is renamed to GN_E2_DEFINE_PUBLIC_ABSTRACT_TYPE.

Also: base/string.h now precedes smartptr.h/rtti.h in GNbase.h so the
mixin can use StrA; auto-generated ids start at 1 so 0 stays an
"invalid/unassigned" sentinel; changed files reformatted with clang-format.

+++ src/core/gpu2/vk-raster-pso-factory.cpp
@@ -189,7 +189,7 @@ Gpu2RasterPsoKey Gpu2RasterPsoKey::make(const GpuShaderVulkan & vs, const GpuSha
 
     // ── 1. Shader identity ────────────────────────────────────────────────────
     // Combine VS and PS ids using FNV multiply to avoid trivial cancellations.
-    k.shaderHash = vs.id * 0x100000001b3ULL ^ (ps ? ps->id : 0ULL);
+    k.shaderHash = uint64_t(vs.id) * 0x100000001b3ULL ^ uint64_t(ps ? ps->id : 0);
 
     // ── 2. Vertex input ───────────────────────────────────────────────────────
     const bool hasInput = !geom.format.attributes.empty() && !geom.vertices.empty();
